fead(MD): add similarity measures

+++ src/core/algorithms/md/md_verifier/thresholds_type.h
@@ -1,5 +1,5 @@
 #include <vector>
 
 namespace algos::md {
-using ThresholdsType = std::vector<double>;
+using ThresholdsType = std::vector<long double>;
 }  // namespace algos::md
